requirements.txt
selenium
pyside6
validator